+++ R/unnest.R
@@ -1,7 +1,8 @@
 #' Unnest a list column.
 #'
 #' If you have a list-column, this makes each element of the list its own
-#' row. List-columns can either be atomic vectors or data frames.
+#' row. `unnest()` can handle list-columns that can atomic vectors, lists, or
+#' data frames (but not a mixture of the different types).
 #'
 #' If you unnest multiple columns, parallel entries must have the same length
 #' or number of rows (if a data frame).
@@ -142,7 +143,7 @@ unnest.data.frame <- function(data, ..., .drop = NA, .id = NULL,
 
 list_col_type <- function(x) {
   is_data_frame <- map_lgl(x, is.data.frame)
-  is_atomic <- map_lgl(x, is_atomic)
+  is_atomic <- map_lgl(x, function(x) !is.object(x) && is_vector(x))
 
   if (all(is_data_frame)) {
     "dataframe"