+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidyr 0.7.2.9000
 
+* `unnest()` can unnested list-columns contains lists of lists (#278).
+
 * `drop_na()` works with list-cols (#280)
 
 * `complete()` with zero-length completions returns original input (#331).